Moved new group button into a menu instead.
