+++ src/main/java/address/controller/PersonOverviewController.java
@@ -107,13 +107,4 @@ public class PersonOverviewController {
         }
         EventManager.getInstance().post(new FilterCommittedEvent(filterExpression));
     }
-
-    @FXML
-    private void handleNewGroup() {
-        ContactGroup tempGroup = new ContactGroup();
-        boolean okClicked = mainController.showGroupEditDialog(tempGroup);
-        if (okClicked) {
-            modelManager.addGroup(tempGroup);
-        }
-    }
 }
